+++ src/main/java/me/clefal/lootbeams/loaders/fabric/LootBeamsFabricModClientEvent.java
@@ -4,21 +4,25 @@
 import me.clefal.lootbeams.LootBeamsConstants;
 
 //? if =1.21.1 {
-/^import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
+import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import me.clefal.lootbeams.compat.fabric_1_21_1.TieredZCompatModule;
 import me.clefal.lootbeams.compat.fabric_1_21_1.TrinketCompatModule;
 import me.clefal.lootbeams.compat.common_1_21_1.AccessoriesCompatModule;
 import me.clefal.lootbeams.compat.common_1_21_1.SubtleEffectCompatModule;
 import me.clefal.lootbeams.compat.common_1_21_1.TieredReforgedCompatModule;
+//? if =1.21.1 && neoforge
 import me.clefal.lootbeams.compat.multiversion_compat.MalumCompatModule;
-^///?}
+//?}
+//? if =1.21.1 {
+
+//?}
 //? if =1.20.1 {
-import me.clefal.lootbeams.compat.fabric_1_20_1.TieredZCompatModule;
+/^import me.clefal.lootbeams.compat.fabric_1_20_1.TieredZCompatModule;
 import me.clefal.lootbeams.compat.fabric_1_20_1.TierifyCompatModule;
 import me.clefal.lootbeams.compat.fabric_1_20_1.TrinketCompatModule;
 import me.clefal.lootbeams.compat.fabric_1_20_1.ZenithCompatModule;
 import me.clefal.lootbeams.compat.multiversion_compat.MalumCompatModule;
-//?}
+^///?}
 
 //? if =1.21.4 {
 /^import me.clefal.lootbeams.compat.common_1_21_4.SubtleEffectCompatModule;
@@ -48,20 +52,21 @@ public class LootBeamsFabricModClientEvent {
         LootBeamsConstants.LOGGER.info("register all modules");
         ModulesManager.registerModules(
                 //? if =1.21.1 {
-                /^SubtleEffectCompatModule.INSTANCE,
+                SubtleEffectCompatModule.INSTANCE,
                 AccessoriesCompatModule.INSTANCE,
                 TieredReforgedCompatModule.INSTANCE,
                 TieredZCompatModule.INSTANCE,
-                TrinketCompatModule.INSTANCE,
-                MalumCompatModule.INSTANCE
-                ^///?}
+                TrinketCompatModule.INSTANCE
+                //? if neoforge
+                ,MalumCompatModule.INSTANCE
+                //?}
                 //? if =1.20.1 {
-                TrinketCompatModule.INSTANCE,
+                /^TrinketCompatModule.INSTANCE,
                 ZenithCompatModule.INSTANCE,
                 TieredZCompatModule.INSTANCE,
                 TierifyCompatModule.INSTANCE,
                 MalumCompatModule.INSTANCE
-                //?}
+                ^///?}
 
                 //? if =1.21.4 {
                 /^SubtleEffectCompatModule.INSTANCE